+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/spinner/DefaultNumberSpinnerBuilder.java
@@ -87,7 +87,7 @@ final class DefaultNumberSpinnerBuilder<T extends Number> extends AbstractSpinne
 	@Override
 	protected ComponentValue<T, JSpinner> createComponentValue(JSpinner component) {
 		if (valueClass.equals(Integer.class) || valueClass.equals(Double.class)) {
-			return new SpinnerNumberValue<>(component);
+			return new SpinnerNumberValue<>(component, valueClass);
 		}
 
 		throw new IllegalStateException("NumberSpinnerBuilder not implemented for type: " + valueClass);